Set fixture remove_document_file on autouse (#132)
* Run fixture remove_document_file automatically on all tests to remove file linked in testcases

* Replaced try/except in fixture with if statement

+++ tests/conftest.py
@@ -1,7 +1,6 @@
 """Shared fixtures and configuration."""
 
 import shutil
-import traceback
 from collections.abc import Generator
 from datetime import datetime
 from datetime import timezone
@@ -43,15 +42,12 @@ def metadata(_mock_timestamp: None) -> DataDocMetadata:
     return DataDocMetadata(str(TEST_PARQUET_FILEPATH))
 
 
-@pytest.fixture()
+@pytest.fixture(autouse=True)
 def remove_document_file() -> Generator[None, None, None]:
     # Yield so we only run teardown
     yield None
-    try:
+    if TEST_RESOURCES_METADATA_DOCUMENT.exists():
         TEST_RESOURCES_METADATA_DOCUMENT.unlink()
-    except FileNotFoundError as e:
-        print("File not deleted on teardown, exception caught:")  # noqa: T201
-        traceback.print_exception(type(e), e)
 
 
 @pytest.fixture()
